Project/requirements/local.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.5
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-import-export==3.0.2
django-js-asset==2.0.0
django-mathfilters==1.0.0
et-xmlfile==1.1.0
filelock==3.9.0
identify==2.5.17
idna==3.4
MarkupPy==1.14
nodeenv==1.7.0
odfpy==1.4.1
openpyxl==3.1.1
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==3.0.4
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
urllib3==1.26.14
virtualenv==20.17.1
xlrd==2.0.1
xlwt==1.3.0
